+++ spec/dummy/test/components/stories/dry_component_stories.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+class DryComponentStories < ViewComponent::Storybook::Stories
+  story :default do
+    constructor(
+      title: text("Hello World!")
+    )
+  end
+end