+++ spec/dummy/test/components/stories/invalid/invalid_constructor_stories.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 module Invalid
-  class InvalidConstrutorStories < ViewComponent::Storybook::Stories
+  class InvalidConstructorStories < ViewComponent::Storybook::Stories
     story :invalid_kwards, ExampleComponent do
       constructor(title: "OK", junk: "Not OK!")
     end